Introduce `control.rs` module
